+++ lib/views/view_accounts/view_accounts_helpers.dart
@@ -39,7 +39,7 @@ extension ViewAccountsHelpers on ViewAccountsState {
                 _selectedPivot[i] = i == index;
               }
               list = getList();
-              selectedItems.clear();
+              selectedItems.value.clear();
             });
           },
           borderRadius: const BorderRadius.all(Radius.circular(8)),